requirements.txt
autopep8==1.6.0
cycler==0.11.0
fonttools==4.33.3
imageio==2.19.3
kiwisolver==1.4.3
matplotlib==3.5.2
networkx==2.8.4
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
pycodestyle==2.8.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.2
scipy==1.8.1
six==1.16.0
tifffile==2022.5.4
toml==0.10.2
